+++ CHANGES.rst
@@ -3,7 +3,8 @@ Change History
 
 0.2 - Unreleased
 ----------------
-- define `IAutowire`.
+- define ``IAutowire``.
+- change ``DoesNotWired`` to subclass of ``AttributeError``
 
 0.1.1 - 2018-06-09
 ------------------